+++ src/fm-client/abilities/external-interface/InventoryAsBatch.js
@@ -1,8 +1,7 @@
 const fs = require('fs');
-const AbstractSinglePage = require('../../components/AbstractSinglePage')
+
+const FileUploadablePage = require('../../components/FileUploadablePage');
 const debug = require('../../../diagnostics/debug')
-const Native = require('../../components/Native');
-const ButtonSymbol = require('../../components/ButtonSymbol');
 const MenuItem = require('../../components/MenuItem')
 
 const CREATE_BUTTON = 2
@@ -11,7 +10,7 @@ const MENU_ITEM = new MenuItem(14, 1, 5)
 /*
  * /外部インターフェース:対HT/棚卸ｲﾝﾎﾟｰﾄ/ 
  */
-module.exports = class InventoryAsBatch extends AbstractSinglePage {
+module.exports = class InventoryAsBatch extends FileUploadablePage {
   constructor(page) {
     super(page)
   }
@@ -20,27 +19,5 @@ module.exports = class InventoryAsBatch extends AbstractSinglePage {
     debug.log('Inventory.enable')
     await super.clickOnMenu(MENU_ITEM, CREATE_BUTTON)
     return true
-  }
-
-  async create(options) {
-    debug.log('Inventory.create')
-    const page = super.page
-    const filePath = options.filePath
-    if(!fs.existsSync(filePath)) return false
-
-    const fileSize = await super.uploadFile(filePath)
-
-    // 登録開始
-    await page.evaluate(Native.performClick(), ButtonSymbol.REGISTER)
-    await super.waitUntilLoadingIsOver()
-    const message = await page.$eval('#errorMessage', e => [e.innerText, e.textContent])
-
-    await page.$eval('.closeBtn', e => e.click())
-
-    return {
-      fileSize: fileSize,
-      isSuccess: message[1].startsWith('インポートに成功しました。'),
-      statusText: message[0]
-    }
   }
 }
